# Fix: Allow same-line comments in padded-blocks (fixes #5055)