--- a/lib/rules/padded-blocks.js
+++ b/lib/rules/padded-blocks.js
@@ -17,20 +17,6 @@ module.exports = function(context) {
         NEVER_MESSAGE = "Block must not be padded by blank lines.";
 
     /**
-     * Retrieves an array of all comments defined inside the given node.
-     * @param {ASTNode} node The AST node.
-     * @returns {ASTNode[]} An array of comment nodes.
-     */
-    function getCommentsInNode(node) {
-        var allComments = context.getAllComments();
-
-        return allComments.filter(function(comment) {
-            return node.range[0] < comment.range[0] &&
-                node.range[1] > comment.range[1];
-        });
-    }
-
-    /**
      * Checks if the location of a node or token is before the location of another node or token
      * @param {ASTNode|Token} a The node or token to check if its location is before b.
      * @param {ASTNode|Token} b The node or token which will be compared with a.
@@ -50,8 +36,15 @@ module.exports = function(context) {
             first = node.body[0],
             firstLine = first.loc.start.line,
             expectedFirstLine = blockStart + 2,
-            comments = getCommentsInNode(node),
-            firstComment = comments[0];
+            leadingComments = (node.body[0].leadingComments || []).slice(),
+            firstComment;
+
+        while (leadingComments.length > 0 &&
+                leadingComments[0].loc.start.line <= node.loc.start.line) {
+            leadingComments.shift();
+        }
+
+        firstComment = leadingComments[0];
 
         if (firstComment && isLocatedBefore(firstComment, first)) {
             firstLine = firstComment.loc.start.line;
@@ -71,8 +64,15 @@ module.exports = function(context) {
             lastToken = context.getLastToken(last),
             lastLine = lastToken.loc.end.line,
             expectedLastLine = blockEnd - 2,
-            comments = getCommentsInNode(node),
-            lastComment = comments[comments.length - 1];
+            trailingComments = (node.body[node.body.length - 1].trailingComments || []).slice(),
+            lastComment;
+
+        while (trailingComments.length > 0 &&
+                trailingComments[trailingComments.length - 1].loc.end.line >= node.loc.end.line) {
+            trailingComments.pop();
+        }
+
+        lastComment = trailingComments[trailingComments.length - 1];
 
         if (lastComment && isLocatedBefore(lastToken, lastComment)) {
             lastLine = lastComment.loc.end.line;