This allows comments that occur on the same line as the opening or closing brace.

By refactoring the `getCommentsInNode` method as part of this change, I was able to throw in a slight (~6%) performance bump. Before:

```
Single File:
  CPU Speed is 2200 with multiplier 13000000
  Performance Run #1:  4078.844035ms
  Performance Run #2:  4145.123632ms
  Performance Run #3:  4228.349988ms
  Performance Run #4:  4022.783567ms
  Performance Run #5:  4196.384683ms

  Performance budget ok:  4145.123632ms (limit: 5909.090909090909ms)


Multi Files (450 files):
  CPU Speed is 2200 with multiplier 39000000
  Performance Run #1:  10260.395154ms
  Performance Run #2:  10054.932538ms
  Performance Run #3:  9980.401873ms
  Performance Run #4:  10205.057365ms
  Performance Run #5:  10123.912088ms

  Performance budget ok:  10123.912088ms (limit: 17727.272727272728ms)
```

After:

```
Single File:
  CPU Speed is 2200 with multiplier 13000000
  Performance Run #1:  3957.185828ms
  Performance Run #2:  3982.306874ms
  Performance Run #3:  3888.848636ms
  Performance Run #4:  3814.4996730000003ms
  Performance Run #5:  3891.909929ms

  Performance budget ok:  3891.909929ms (limit: 5909.090909090909ms)


Multi Files (450 files):
  CPU Speed is 2200 with multiplier 39000000
  Performance Run #1:  10070.100405ms
  Performance Run #2:  10050.302239ms
  Performance Run #3:  10265.260417ms
  Performance Run #4:  10013.769042ms
  Performance Run #5:  10078.269855ms

  Performance budget ok:  10070.100405ms (limit: 17727.272727272728ms)
```