@lo1tuma 2016-02-01
`getCommentsInNode` [was introduced as a workaround](https://github.com/eslint/eslint/pull/2379) for https://github.com/eslint/espree/issues/137 which is fixed now. So maybe we can get rid of it completely which should also improve the performance.