@nzakas 2016-02-01
Ooh nice. @btmills do you want to try removing `getCommentsInNode()`?