@btmills 2016-02-06
:bomb: :boom: `getCommentsInNode()` is gone!

Master:

```
Single File:
  CPU Speed is 2200 with multiplier 13000000
  Performance Run #1:  5298.615065ms
  Performance Run #2:  5634.129176ms
  Performance Run #3:  5402.509475ms
  Performance Run #4:  5322.369708ms
  Performance Run #5:  5269.647078ms

  Performance budget ok:  5322.369708ms (limit: 5909.090909090909ms)
```

This PR:

```
Single File:
  CPU Speed is 2200 with multiplier 13000000
  Performance Run #1:  4939.508521ms
  Performance Run #2:  5011.416262ms
  Performance Run #3:  4993.326752ms
  Performance Run #4:  4963.628394ms
  Performance Run #5:  4929.379918ms

  Performance budget ok:  4963.628394ms (limit: 5909.090909090909ms)
```